requirements.txt
ultralytics==8.0.170
deep_sort_realtime==1.3.1
streamlit==1.35.0
mlflow==2.13.0
torch==2.5.1
torchvision==0.20.0
opencv-python
numpy==1.23.5
pandas
scikit-learn
matplotlib
seaborn
tqdm
pillow

